move js loader to bundler pkg
