+++ packages/docusaurus-bundler/package.json
@@ -18,13 +18,25 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@babel/core": "^7.23.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.22.9",
+    "@babel/preset-env": "^7.22.9",
+    "@babel/preset-react": "^7.22.5",
+    "@babel/preset-typescript": "^7.22.5",
+    "@babel/runtime": "^7.22.6",
+    "@babel/runtime-corejs3": "^7.22.6",
     "@docusaurus/cssnano-preset": "3.5.2",
     "@docusaurus/faster": "3.5.2",
     "@docusaurus/logger": "3.5.2",
     "@docusaurus/types": "3.5.2",
+    "@docusaurus/utils": "3.5.2",
+    "babel-loader": "^9.1.3",
+    "babel-plugin-dynamic-import-node": "^2.3.3",
     "clean-css": "^5.3.2",
     "copy-webpack-plugin": "^11.0.0",
     "css-minimizer-webpack-plugin": "^5.0.1",
+    "fs-extra": "^11.1.1",
     "mini-css-extract-plugin": "^2.9.1",
     "react-dev-utils": "^12.0.1",
     "terser-webpack-plugin": "^5.3.9",
