+++ packages/docusaurus-bundler/src/importFaster.ts
