+++ packages/docusaurus-bundler/src/loaders/babelPreset.ts
