+++ packages/docusaurus/package.json
@@ -35,13 +35,6 @@
   "dependencies": {
     "@babel/core": "^7.23.3",
     "@babel/generator": "^7.23.3",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.22.9",
-    "@babel/preset-env": "^7.22.9",
-    "@babel/preset-react": "^7.22.5",
-    "@babel/preset-typescript": "^7.22.5",
-    "@babel/runtime": "^7.22.6",
-    "@babel/runtime-corejs3": "^7.22.6",
     "@babel/traverse": "^7.22.8",
     "@docusaurus/bundler": "3.5.2",
     "@docusaurus/logger": "3.5.2",
@@ -50,8 +43,6 @@
     "@docusaurus/utils-common": "3.5.2",
     "@docusaurus/utils-validation": "3.5.2",
     "autoprefixer": "^10.4.14",
-    "babel-loader": "^9.1.3",
-    "babel-plugin-dynamic-import-node": "^2.3.3",
     "boxen": "^6.2.1",
     "chalk": "^4.1.2",
     "chokidar": "^3.5.3",
